factored out one more commonality to list.js
